+++ packages/cryptoplease/lib/app/screens/authenticated/wallet_flow/wallet_flow_screen.dart
@@ -7,7 +7,7 @@ import '../../../../core/currency.dart';
 import '../../../../core/presentation/format_amount.dart';
 import '../../../../core/tokens/token_list.dart';
 import '../../../../di.dart';
-import '../../../../features/incoming_tip_payments/module.dart';
+import '../../../../features/incoming_tip_payments/process_incoming_tip.dart';
 import '../../../../features/outgoing_direct_payments/module.dart';
 import '../../../../features/outgoing_split_key_payments/module.dart';
 import '../../../../features/outgoing_tip_payments/module.dart';
@@ -43,7 +43,7 @@ class _State extends State<WalletFlowScreen> {
     if (!mounted) return;
 
     if (request is QrScannerTipRequest) {
-      await context.processIT(request.tipData);
+      await context.processIncomingTip(request.tipData);
 
       return;
     }
