+++ packages/cryptoplease/lib/data/db/db.dart
@@ -6,7 +6,7 @@ import '../../features/activities/db.dart';
 import '../../features/activities/models/transaction.dart';
 import '../../features/favorite_tokens/db.dart';
 import '../../features/incoming_split_key_payments/db.dart';
-import '../../features/incoming_tip_payments/module.dart';
+import '../../features/incoming_tip_payments/db.dart';
 import '../../features/outgoing_direct_payments/module.dart';
 import '../../features/outgoing_split_key_payments/module.dart';
 import '../../features/outgoing_tip_payments/module.dart';
