+++ packages/cryptoplease/lib/features/incoming_tip_payments/db.dart
@@ -1,0 +1,1 @@
+export 'src/bl/it_repository.dart' show ITRows, ITStatusDto;
