+++ packages/cryptoplease/lib/features/incoming_tip_payments/routes.dart
@@ -1,6 +1,6 @@
 import 'package:auto_route/auto_route.dart';
 
-import 'src/presentation/incoming_tip_screen.dart';
+import 'src/widgets/incoming_tip_screen.dart';
 
 const incomingTipRoutes = [
   AutoRoute<void>(page: IncomingTipScreen),
